Position: use BigDecimal instead of double
for cost, realizedPL, marketPrice, marketValue, averagePrice,
unrealizedPL and exposure

+++ common/src/main/java/ch/algotrader/simulation/Simulator.java
@@ -387,7 +387,7 @@ public class Simulator {
         double amount = 0.0;
         for (Position openPosition : openPositions) {
             final MarketDataEventVO marketDataEvent = this.marketDataCache.getCurrentMarketDataEvent(openPosition.getSecurity().getId());
-            amount += openPosition.getMarketValue(marketDataEvent);
+            amount += openPosition.getMarketValue(marketDataEvent).doubleValue();
         }
         return amount;
     }
